pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "djogenes"
description = 'Django application to manage your diogenes syndrome (too much stuff in house)'
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "lasizoillo", email = "lasizoillo@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "Django~=4.1",
  "django-environ==0.9.0",
  "psycopg2==2.9.5",
  "django-htmx==1.13.0",
  "django-bulma==0.8.3",
  "django-pipeline==2.0.9",
  "django-soft-delete==0.9.21",
  "django-admirarchy==1.2.2",
  "Pillow==9.4.0"
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/unknown/djogenes#readme"
Issues = "https://github.com/unknown/djogenes/issues"
Source = "https://github.com/unknown/djogenes"

[tool.hatch.version]
path = "djogenes/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=djogenes --cov=tests {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "djogenes/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
